chore: Relocate customDurationResolver to MainActivity.kt
- Remove redundant init block from onCreate
