+++ app/src/main/java/ke/don/koffee_demo/KoffeeDemoApplication.kt
@@ -10,25 +10,5 @@
 package ke.don.koffee_demo
 
 import android.app.Application
-import ke.don.koffee.domain.Koffee
-import ke.don.koffee.model.ToastDuration
-import ke.don.koffee.ui.DefaultToast
 
-class KoffeeDemoApplication : Application() {
-    override fun onCreate() {
-        super.onCreate()
-
-        Koffee.init {
-            layout { DefaultToast(it) } // or custom
-            dismissible(true)
-            durationResolver { customDurationResolver(it) }
-        }
-    }
-}
-
-fun customDurationResolver(duration: ToastDuration): Long? = when (duration) {
-    ToastDuration.Short -> 5000L
-    ToastDuration.Medium -> 7000L
-    ToastDuration.Long -> 10000L
-    ToastDuration.Indefinite -> null
-}
+class KoffeeDemoApplication : Application()
